Moved common functionality to commons
Functions moved:
flatten
num_to_let_dict
let_to_num_dict
num_to_let
canonicalize_indices
is_canonical
